refactor(engine): relocate SandboxRunner abstraction from nebula-sandbox
The runner abstraction belongs with the dispatcher that owns it.
`SandboxRunner`, `SandboxedContext`, `InProcessSandbox`,
`ActionExecutor`/`ActionExecutorFuture`, the `impl SandboxRunner for
ProcessSandbox` adapter, and `sandbox_error_to_action_error` move into
`nebula-engine` (`runtime::sandbox_runner`). The engine import sites and
the `pub use` re-export are repointed in this commit — no compatibility
shim.

`ProcessSandboxHandler` is decoupled from the `SandboxRunner` trait: it
now calls the transport `ProcessSandbox::invoke_with_cancel` directly and
maps `SandboxError` -> `ActionError` at the `StatelessHandler` boundary
(so it stays compilable while the trait lives in the engine and before
the discovery cluster relocates to nebula-plugin). The `SandboxError` ->
`ActionError` classification tests move with the function.
