+++ crates/sandbox/src/handler.rs
@@ -8,7 +8,7 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use nebula_action::{ActionContext, ActionError, ActionMetadata, ActionResult, StatelessHandler};
 
-use crate::{SandboxRunner, dispatch::ProcessSandbox, runner::SandboxedContext};
+use crate::{dispatch::ProcessSandbox, error::sandbox_error_to_action_error};
 
 /// Wraps a [`ProcessSandbox`] as a [`StatelessHandler`].
 ///
@@ -16,7 +16,8 @@ use crate::{SandboxRunner, dispatch::ProcessSandbox, runner::SandboxedContext};
 /// When the engine calls `execute()`, the request is routed through the
 /// sandbox's long-lived plugin process using the duplex envelope transport
 /// (ADR 0006; handshake + dialed socket), not direct stdin/stdout action
-/// invocation.
+/// invocation. The transport returns a `SandboxError`; this handler maps it
+/// to the `ActionError` taxonomy at the `StatelessHandler` boundary.
 pub struct ProcessSandboxHandler {
     sandbox: Arc<ProcessSandbox>,
     metadata: ActionMetadata,
@@ -40,7 +41,19 @@ impl StatelessHandler for ProcessSandboxHandler {
         input: serde_json::Value,
         context: &dyn ActionContext,
     ) -> Result<ActionResult<serde_json::Value>, ActionError> {
-        let sandboxed = SandboxedContext::new(context);
-        self.sandbox.execute(sandboxed, &self.metadata, input).await
+        // Call the transport directly and map its `SandboxError` to the
+        // engine's `ActionError` taxonomy here. The transport crate does
+        // not know about `ActionError`; the round-trip is raced against
+        // the action's cancellation token (a pre-cancelled token resolves
+        // on the first poll, so no separate pre-flight check is needed).
+        self.sandbox
+            .invoke_with_cancel(
+                self.metadata.base.key.as_str(),
+                input,
+                context.cancellation(),
+            )
+            .await
+            .map(ActionResult::success)
+            .map_err(sandbox_error_to_action_error)
     }
 }
